+++ packages/conc/changelog.md
@@ -3,6 +3,7 @@
 * Add `lock`, a combinator for protecting a region with a mutex.
 * Add `scheduleAsync`, a combinator for running an action async that allows the handle to be used before the thread
   starts
+* Change the default signal handler for `Interrupt` to `CatchInfo`, catching repeated signals.
 
 # 0.3.0.0